Changed to fit repository structure
Changed name of LIC Test file to LICTest.

Moved LIC5 to EveluateLIC.
